Reorganize GC-related submodules in `wasmtime`
