+++ crates/wasmtime/src/runtime/gc/disabled.rs
