+++ crates/wasmtime/src/runtime/gc/enabled.rs
