+++ crates/wasmtime/src/runtime/gc/enabled/rooting.rs
