+++ crates/wasmtime/src/runtime/store.rs
@@ -79,7 +79,7 @@
 use crate::instance::InstanceData;
 use crate::linker::Definition;
 use crate::module::{BareModuleInfo, RegisteredModuleId};
-use crate::r#ref::RootSet;
+use crate::gc::RootSet;
 use crate::trampoline::VMHostGlobalContext;
 use crate::{module::ModuleRegistry, Engine, Module, Trap, Val, ValRaw};
 use crate::{Global, Instance, Memory};
